Remove allow(unused_variables) from generated code.
This commit removes allow(unused_variables) from the generated code.

We now camel case the binding names from the Rust identifiers and ensure two
parameters don't camel case to the same value.

This enables developers to receive the unused variables warnings and to solve
the warnings (i.e. prefix the variable name with `_`) in the suggested,
Rust-idiomatic way.

Fixes #6.