Repository refactoring.
This major refactoring is an attempt to share the definitions of the codegen
bindings between the `azure-functions-codegen` and `azure-functions` crates.

Two new crates were added:

* `azure-functions-shared` to share the code between the above two crates.
* `azure-functions-shared-codegen` to facilitate the import of code that was
generated from protobuf definitions.

The protobuf submodule now exists under the `azure-functions-shared` crate.

Initially, I attempted to have `azure-functions-shared` also be responsible for
implementing the `TryFrom<AttributeArguments>` and `IntoToken` traits for the
bindings, but this resulted in the crate linking against compiler libraries at
runtime. For now, the code for that still lives in `azure-functions-codegen`.

The bindings now have a single source of truth using `Cow` types (stores static
information for the generated code and owned information during procedural
macro evaluation).  This should help make it a little easier to define new
bindings going forward.