Add tests for the azure-functions crate.
This commit adds a number of tests to improve test coverage for the
azure-functions crate.

A bug was caught during testing where `HttpResponse` was not properly setting
the status code in the response; this was due to the recent refactoring where
another `From` implementation was added and the existing `From` implementation
(where the status code was being set) was no longer being invoked.

The documentation examples have also been improved with this commit.