Implement input/output blob bindings.
This commit implements the `azure_functions::bindings::Blob` type that is
responsible for input and output blob bindings.

Note: the inout binding for Blob is not yet implemented because some work needs
to be done to implement inout bindings (the current mechanism is insufficient
for types that support `in`, `inout`, and `out` like blob bindings do).