Implement 'inout' bindings.
This commit refactors code generation to support `inout` direction bindings.

Parameters passed as `&mut` are automatically inferred to be `inout` direction.

This commit also changes how data is passed to the invoker so that input data
can be moved directly into the bindings.  As a result, the API for blob and
queue bindings has been simplified.

Fixes #25.