Implement initial SDK executable.
This commit implements `cargo func new-app` to help create a new Azure
Functions project.

It creates the crate, adds the dependencies for Azure Functions for Rust,
initializes a main.rs with the expected content, and creates Docker files that
will be used for future commands for building and running the functions.