Fix build for recent nightly change.
This commit fixes the build break introduced by the latest nightly compiler:
the 'underscore' lifetime cannot be used in the context of a type parameter
constraint.

The cargo dependencies are also updated and the example Dockerfile has been
updated to use the final Azure Functions 2.0 base image.

Additionally, the source has been reformatted for the latest rustfmt.