Implement `sync-extensions` worker command.
This commit implements a `sync-extensions` command for the language worker.
The command is responsible for generating a .NET Core project file that can be
used to restore and publish the required Azure Functions binding extensions.

Currently, only one binding extension is supported, which is for storage.  If
you use storage bindings the `sync-extensions` command will automatically
restore the binding extensions for the Azure Functions storage bindings.

Fixes #53.