Refactor the `register` macro to `export`.
This refactoring enables a project layout that will facilitate a cargo command
for adding new Azure Functions to an Azure Functions for Rust project.

This renames the `register` macro to `export` (an expression macro).

Additionally, the macro no longer generates a main function.  The documentation
and the template created by `cargo func new-app` have been updated.