Use OpenSSL for grpcio crate.
This commit changes the grpcio crate to build against OpenSSL instead of
BoringSSL.

Doing so enables Azure Functions for Rust to be used with crates that also link
against OpenSSL, such as reqwest crate.

This also removes Golang as a dependency for building Azure Functions for Rust.

Also disabling clippy linting from the CI for now.

Fixes #108.