Remove mod.rs files.
Removing `mod.rs` files where appropriate.

Unfortunately, we can't remove mod.rs from the examples or from
`cargo func new-app` because there appears to be a bug in the Rust compiler
where a procedural macro emits a mod statement and the modules are not in the
same directory as the file (i.e. `functions.rs` vs. `functions/mod.rs`).