Remove Trigger trait.
This commit removes the `Trigger` trait that was used to read the request
metadata into the trigger binding.

Instead, trigger types now implement a `new` method that takes both the bound
parameter data and the trigger metadata.

This eliminates the need for the trigger types to use `Option` types for
fields because the initialization happens in one step rather than two.

Additionally, `#[doc(hidden)]` was applied to anything that takes the internal
`TypedData` to hide it from the documentation.