Merge the binding types.
This commit merges the binding types from the codegen crate with those that are
from the shared crate.

The bindings in the codgen crate were responsible for parsing `binding`
attributes and generating code to define the bindings.

The bindings in the shared crate were responsible for serializing the data to
function.json.

With a dependency on `proc_macro2`, we can have a single type definition that
is both responsible for the code generation and for the JSON serialization.

The combined types are not implemented in the shared crate.