Fix serialization of empty array values for bindings.
This commit fixes the serialization of empty array values for bindings into
`function.json`.

Currently this only impacts the HTTP trigger's `methods` attribute as it is the
only string array.  Without this fix, an empty array gets serialized and HTTP
functions cannot be triggered because no verb is accepted by the Azure
Functions Host.

Additionally, the validation performed to string values and string array values
was case sensitive.  This is also addressed to ensure the comparison is done in
the correct case.

Fixes #187.