Remove mutex around registry.
This commit removes the mutex that was guarding access to the function
registry.

Additionally, only invocation requests are moved to the thread pool.  Thus, all
other messages are handled by the main thread of execution and can quickly
interact with the function registry as needed without having to be locked.