Pass trigger and input bindings by value.
This commit changes the way trigger and input bindings are passed to Azure
Functions.

Previously, trigger and input bindings were passed by immutable reference.
However, if the function wanted to output the input data, it would have to
clone the input data, creating an unnecessary copy.

Now both trigger and input bindings are passed by value.  The input data can
now be moved to the output data as necessary.