Change `CosmosDbDocuments` to `CosmosDbDocument`.
This commit changes `CosmosDbDocuments` to the singular `CosmosDbDocument` and
changes the API to treat the binding as 1:1 with a JSON document object.

The intention with this change is to make this API more consistent with the
other input and output binding types that are singular in nature.

In an upcoming change, I will add support for `Vec<T>`, where T is a binding
type, so that multiple input and output items for a single binding will be
supported.