Fix build Docker image and update README.
This commit fixes the build Docker image to include a .NET Core SDK
install.  This allows extension sync to occur since it depends on a .NET
Core SDK being installed.

Additionally, the `Dockerfile` template has been updated to use the
`runtime-deps` as the base image.  This saves ~150 MB of space in the
resulting deployment image.

The README has been updated to call out all dependencies, including a C++
compiler and a .NET Core SDK.

Fixes #231.