Implement copying existing local settings file.
This commit implements copying an existing local settings file to the script
root during application initialization.

It adds the `--local-settings` option to manually specify the settings file to
use.

If not specified and the `init` command is running from `cargo`, it will
default to the local settings file in the manifest directory.

If a local settings file does not exist at the default location, then a default
one will be created in the script root.