Replace grpc-rs with tower-grpc.
This commit replaces the grpc-rs dependency with tower-grpc.

As a result, this also replaces the rust-protobuf dependency with prost.

Prost generates far less code that has the benefit of being simple structs and
enums.

Because tower-grpc is entirely Rust native, we no longer need to depend on a
CMake or C++ compiler to be installed to build Azure Functions for Rust.

Fixes #243.
Fixes #220.