Implement generic trigger, input, and output bindings.
This commit implements generic trigger, input, and output bindings that enable
the integration of unsupported binding extensions.

The generic bindings will export whatever attribute values are given in the
binding attributes to function.json.

Additionally, there is a simple representation (a `generic::Value`) of the gRPC
data that is available in the generic bindings.

This includes an example that is a port of the CosmosDB example, except none of
the supported CosmosDB bindings are used.

Fixed a bug where using a CosmosDB document (but not the CosmosDB trigger)
binding failed to synchronize the CosmosDB extension.

Closes #34.