Implement async Azure Functions.
This commit implements support for async Azure Functions.

The feature is behind the `unstable` feature for the azure-functions crate.

Additionally, users must have a dependency on `futures-preview` for the
generated code to build.

Part of this work was refactoring the worker code out of the run command and
into its own file.

The thread-local invocation context also needed to be refactored to better
support a futures-based invocation. This is a breaking change as `Context` is
no longer a binding type; instead, users use `Context::current` to get the
current invocation context.

Closes #270.