Remove export! macro.
This commit removes the `export!` macro from `azure-functions`.

The macro was previously used to declare the given modules as well as generate
a `FUNCTIONS` constant passed to the main function.

Unfortunately because the modules were being defined dynamically, syntax errors
in the function source files would mean a completely undiagnosable error
message would be emitted by the compiler pointing at the `export!` macro.

The fix is to remove the `export!` macro entirely and implement a more
sophisticated `cargo func new` that relies on no macro implementations.

Fixes #335.