Additional Durable Functions implementation.
This commit continues the Durable Functions implementation:

* Changes `azure-functions-durable` crate to use the released version of the
`hyper` crate, but with the 0.1.0 futures-compatibility layer.  This fixes the
issue where the reactor could not be found for the client with a 0.1.0
futures-based tokio reactor that is being used by the Azure Functions for Rust
runtime.

* Implement the `DurableOrchestrationClient` binding to wrap an
`OrchestrationClient` and provide the `start_new` method.  More methods will be
implemented later.

* Implement the `DurableActivityContext` binding to extract the instance
identifier and input from the binding data.

* Implemented a mechanism for outer orchestration futures to control inner
futures for updating the current orchestration event, which should occur
immediately after the future is ready.

* Implemented proper replay detection.

* Implemented the `join_all` method on `DurableOrchestrationContext`.

* Implemented the Durable Functions example (still needs documentation).

* Various clippy and code clean up fixes.