Implement `DurableOrchestrationContext::select_all`.
This commit implements the `select_all` function on
`DurableOrchestrationContext`.

The function can be used to select from multiple futures, returning the ready
future's output, the index from the original sequence, and the remaining
pending futures.

Additionally, the orchestration state now properly moves to the correct
orchestration started / completed events when handling the resolution of a
future.