Implement DurableOrchestrationContext::call_activity_with_retry.
This commit implements the `call_activity_with_retry` method.

It refactors the code to return `ActionFuture` rather than `impl Future`.  This
will allow the `join_all` and `select_all` methods to use a mix of
`call_activity` and `call_activity_with_retry`, which is not possible if both
methods return `impl Future`.