Refactor HttpRequest and HttpResponse.
This commit refactors HttpRequest and HttpResponse to directly store the data
rather than an inner `RpcHttp`.

It also implements support for cookies in `HttpRequest`, `HttpResponse`, and
`ResponseBuilder`.

Also enables more lints and fixes the clippy warnings.

Binding attributes for parameters have moved to the parameters in the examples
and documentation.

Fixes #346.
Fixes #281.